Moved useColor var closer to top of the file as it was referenced before the definition if debug=1
Changed execution of $cmd to eval "$cmd" as commands that set a var on the CL before the script in the JSON file would not exe properly.

ie. RUNTIME_FATAL_WARNINGS=1 ./runcheck.sh ...

That would exit with command not found

simulated on the command line:
"x=1 ls" - does not work
eval "x=1 ls" - this does work